More updates buildignore
This has all been kind of weird but basically I think I finally got everything pertaining to the parallelprep function to be excised from the built package including the .Rd documentation file which was delted by document(). The trick was to run devtools::document() after making the correct changes to the .Rbuildignore.